django-app/requirements/local.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.12.0
django-environ==0.8.1
django-filter==21.1
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.1
environ==1.0
h11==0.13.0
idna==3.3
inflection==0.5.1
jsonschema==4.5.1
Markdown==3.3.7
oauthlib==3.2.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
whitenoise==6.1.0
uvicorn[standard]==0.17.6
celery==5.2.1
flower==1.0.0
psycopg2-binary==2.9.2
redis==4.0.2
django-celery-beat==2.2.1
django-redis==5.2.0  # https://github.com/jazzband/django-redis
channels==3.0.4
channels_redis==3.4.0